requirements.txt
numpy
scipy
torch
transformers
soundfile
pydub
bittensor
openai-whisper
pytest